+++ repository-module/src/main/java/by/tut/mdcatalog/project2/repository/GenericRepository.java
@@ -1,0 +1,8 @@
+package by.tut.mdcatalog.project2.repository;
+
+import java.sql.Connection;
+
+public interface GenericRepository {
+
+    Connection getConnection();
+}
