--- repository-module/src/main/java/by/tut/mdcatalog/project2/repository/connection/ConnectionService.java
@@ -1,8 +1,0 @@
-package by.tut.mdcatalog.project2.repository.connection;
-
-import java.sql.Connection;
-
-public interface ConnectionService {
-
-    Connection getConnection();
-}
