--- repository-module/src/main/java/by/tut/mdcatalog/project2/repository/properties/DatabaseProperties.java
@@ -1,39 +1,0 @@
-package by.tut.mdcatalog.project2.repository.properties;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-public class DatabaseProperties {
-
-    @Value("${spring.datasource.url}")
-    private String url;
-    @Value("${spring.datasource.username}")
-    private String username;
-    @Value("${spring.datasource.password}")
-    private String password;
-    @Value("${spring.datasource.driver-class-name}")
-    private String driver;
-    @Value("${database.init.file}")
-    private String initFile;
-
-    public String getUrl() {
-        return url;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String getDriver() {
-        return driver;
-    }
-
-    public String getInitFile() {
-        return initFile;
-    }
-}
